# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Entwine Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  _id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  images: string[];
  category: string;
  /** @nullable */
  fabric?: string | null;
  sizes?: string[];
  inStock: boolean;
  featured?: boolean;
  onSale?: boolean;
  tags?: string[];
  createdAt?: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  pages: number;
}

export interface Category {
  _id: string;
  name: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  description?: string | null;
  productCount?: number;
}

export interface CartItem {
  _id: string;
  productId: string;
  name: string;
  price: number;
  image: string;
  quantity: number;
  size: string;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: string;
  quantity: number;
  size: string;
}

export interface CartItemUpdate {
  sessionId: string;
  quantity: number;
}

export interface CartSessionRef {
  sessionId: string;
}

export interface StoreSummary {
  totalProducts: number;
  totalCategories: number;
  saleCount: number;
  featuredCount: number;
}

export type GetProductsParams = {
category?: string;
search?: string;
onSale?: boolean;
featured?: boolean;
page?: number;
limit?: number;
};

export type GetCategoryProductsParams = {
page?: number;
limit?: number;
};

export type GetCartParams = {
sessionId: string;
};