# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Entwine Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: onSale
          in: query
          schema:
            type: boolean
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products for homepage carousel
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/sale:
    get:
      operationId: getSaleProducts
      tags: [products]
      summary: Get products on sale
      responses:
        "200":
          description: Sale products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /categories/{slug}/products:
    get:
      operationId: getCategoryProducts
      tags: [categories]
      summary: Get products by category slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Products in category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartSessionRef"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /store/summary:
    get:
      operationId: getStoreSummary
      tags: [products]
      summary: Get store summary stats
      responses:
        "200":
          description: Store summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required:
        - _id
        - name
        - slug
        - price
        - images
        - category
        - inStock
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        category:
          type: string
        fabric:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        onSale:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string

    ProductsResponse:
      type: object
      required:
        - products
        - total
        - page
        - pages
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer

    Category:
      type: object
      required:
        - _id
        - name
        - slug
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        image:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        productCount:
          type: integer

    CartItem:
      type: object
      required:
        - _id
        - productId
        - name
        - price
        - image
        - quantity
        - size
      properties:
        _id:
          type: string
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        image:
          type: string
        quantity:
          type: integer
        size:
          type: string

    Cart:
      type: object
      required:
        - sessionId
        - items
        - subtotal
        - itemCount
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        itemCount:
          type: integer

    CartItemInput:
      type: object
      required:
        - sessionId
        - productId
        - quantity
        - size
      properties:
        sessionId:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        size:
          type: string

    CartItemUpdate:
      type: object
      required:
        - sessionId
        - quantity
      properties:
        sessionId:
          type: string
        quantity:
          type: integer

    CartSessionRef:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string

    StoreSummary:
      type: object
      required:
        - totalProducts
        - totalCategories
        - saleCount
        - featuredCount
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        saleCount:
          type: integer
        featuredCount:
          type: integer